feat: combohash highlighter for cyber security professionals
The combohash highlight is a simple highlighter for penetration testers. For files in the format of 'username:password/hash # Comment' it will do higlighting.
This is valuable for both penetration testers and their clients, as often work with an ever-growing list of credentials as we work our way through an enviroment during a penetration test.